feat!: Add parameter binding to operation execution (#15)
The method `dbapi2.Cursor.execute() supports now parameter binding.
Parameters may be provided as a mapping and will be bound to variables in the operation.
In the current version, the parameter style `named` only is implemented.

BREAKING CHANGE
The `parameters` argument in the `Cursor.execute()` method now requires the presence of
two keys:
- `"payload"` where the old payload object is stored, now placeholder variables must replace the values
- `"params"` where the mapping for the variables is stored, here each variable is associated with the value to  be
supplied.

Closes #15