refactor: Refactor SQL `INSERT` statement (#55)
The class `Insert` has been completely redesigned with the `.values()` and `.returning()` features
for the corresponding statement clauses.
Now, the class `Insert` uses the new `Table` class to operate on row columns. To this end, the `Table` class
has been updated to provide an `.insert()` constructor method too. This takes care of setting the parent
object for the newly created `Insert` object propertly.
The `.insert()`constructor has been updated as well accordingly.

- Add the `.values() feature
- Add the `.returning()` feature
- Update `Table` to provide an `.insert()` method constructor
- Fix issue related to Notion specific columns: The parent object was not set, so the columns did not belong to any
table.
- Update/add tests and docs