feat(sql): Add `not_in`, `endswith` for `String` and `before` for `Date`.
This version adds 2 new string operators and the first date operator.
Additionally, it introduces the `CompilerState` data structure for holding important
compilation information. This new dataclass is applied to the compilation of binary expressions.
The `Date` filter properties pose a new challenge: How to compile the values for late binding?
The `_compile_type_filter()` now handles bind parameters and it generates place holders.
The actual values are stored as `BindParameters` in the compiler state's `binds` class attribute.
At execution time, the `ExecutionContext` is responsible to substitute the placeholders with
the effective values.

Relates to #140.